+++ lib/src/widgets/widgets.dart
@@ -6,6 +6,7 @@ export 'header_bar.dart';
 export 'header_button.dart';
 export 'popup_menu.dart';
 export 'sidebar.dart';
+export 'stack_sidebar.dart';
 export 'preferences_group.dart';
 export 'view_switcher.dart';
 export 'view_switcher_tab.dart';