+++ CHANGELOG.md
@@ -4,5 +4,10 @@
 - Rename Every widget from GtkSomething to AdwSomething
 - Rename GtkContainer to AdwClamp and GtkTwoPane to AdwFlap
 - Add `AdwPreferenceGroup` and `AdwActionRow` from libadwaita.
+- Add `AdwStackSidebar` which is basically `GtkStackSidebar`
+- `AdwHeaderBar` parameter's
+    - Replace leading with start
+    - Replace trailing with end
+    - Replace center with title
 
 For older Changelog visit: https://pub.dev/packages/gtk/changelog